LPS-146691 Creates new implementations for InfoItemObjectVariationProvider for Journal and DL
